+++ aol/wsgi.py
@@ -6,8 +6,6 @@ import sys
 def create_wsgi_application(root, settings_module=None, local_settings_file=None):
     """Create a WSGI application anchored at ``root``.
 
-    ``root`` must contain a virtualenv at ``./.env``.
-
     ``settings_module`` is only used if the ``DJANGO_SETTINGS_MODULE``
     environment variable is not already set.
 
@@ -19,22 +17,6 @@ def create_wsgi_application(root, settings_module=None, local_settings_file=None
     """
     if local_settings_file is None:
         local_settings_file = os.path.join(root, 'local.cfg')
-
-    major, minor = sys.version_info[:2]
-    site_packages = 'lib/python{major}.{minor}/site-packages'.format(**locals())
-    site_packages = os.path.join(root, '.env', site_packages)
-
-    if not os.path.isdir(site_packages):
-        raise NotADirectoryError(
-            'Could not find virtualenv site-packages at {}'.format(site_packages))
-
-    # Add the virtualenv's site-packages to sys.path, ensuring its packages
-    # take precedence over system packages (by moving them to the front of
-    # sys.path after they're added).
-    old_sys_path = list(sys.path)
-    site.addsitedir(site_packages)
-    new_sys_path = [item for item in sys.path if item not in old_sys_path]
-    sys.path = new_sys_path + old_sys_path
 
     os.environ.setdefault('LOCAL_SETTINGS_FILE', local_settings_file)
     os.environ.setdefault('DJANGO_SETTINGS_MODULE', settings_module)